Update LED status file path and existence checks
Changed the LED status file path from the user's home directory to '/expidite/tmp/tmp_flags/led_status' across bcli.py, device_manager.py, and led_control.py. Improved file existence checks before reading or writing to the status file to prevent errors if the directory does not exist.

+++ src/expidite_rpi/core/device_manager.py
@@ -14,7 +14,7 @@ from expidite_rpi.utils import utils
 
 logger = root_cfg.setup_logger("expidite")
 
-LED_STATUS_FILE: Path = Path.home() / ".expidite" / "flags" / "led_status"
+LED_STATUS_FILE = Path("/expidite") / "tmp" / "tmp_flags" / "led_status"
 
 class DeviceManager:
     """Manages LED & Wifi status if configured to do so in my_device (DeviceCfg):
@@ -138,9 +138,9 @@ class DeviceManager:
         - status: "on", "off", or "blink"
         """
         try:
-            LED_STATUS_FILE.parent.mkdir(parents=True, exist_ok=True)
-            with open(LED_STATUS_FILE, "w") as f:
-                f.write(f"{colour}:{status}\n")
+            if LED_STATUS_FILE.parent.exists():
+                with open(LED_STATUS_FILE, "w") as f:
+                    f.write(f"{colour}:{status}\n")
         except Exception as e:
             logger.error(f"{root_cfg.RAISE_WARN()}set_led_status threw an exception: " + str(e), 
                          exc_info=True)
